Create base MXTests class, derive MX1Tests and MX2Tests
